Remove neural-network-training realignment entirely.
Add a "delta" test: wherein the bird's chirp is a delta function in time, and the canonical signal is the same.

Change everything to the normalisation described in the paper--zscore for both columns (zscore(nnsetX)) and rows (network's norm function is 'mapstd').

Remove unnecessary convenience variables from the network object savefile, since they're redundant--already in "net".