New code to look at syllable transition probabilities!
A little rough around the edges ;)

Give it a network.  It assumes that it can find the files in /Volumes/Data/song/ BIRD / * /chop_data/wav/, and reads as many as it can, calling out to Nathan's offline syllable detector for nn processing.  Then it produces state transition maps for every hour.  That's all for now.